2023 data changes (#3)
The data received from BoM in October 2023 did not include wet bulb temperature, relative humidity or standard deviation of wind direction. This update removes those variables from the processing.

Relative humidity is calculated inline from temperature and dew point.

Wet bulb is not calculated, as the `metpy.calc.wet_bulb_temperature` function fails if any of the input values are `nan` or `inf`, and I didn't want to fill the missing values.

Also removes a bug where the event data was written to a pickle file, irrespective of what configuration settings are used.