Join weather codes (#6)
* Remove stray breakpoint call
* Add a check for stray breakpoint calls
* WIP: Updating the classifier
* Fix long line in log message
* Add provenance to convert storm count data
Also clean up the classifyTimeSeries to make variable names a bit more intuitive

* Extract accumulated rainfall
Extract accumulated precipitation in the hour preceding and the hour following the daily maximum gust event. Exclude any rainfall observations that have not passed BoM quality checks (i.e. only include obs where the rainfall quality flag is 'Y').

* Remove duplicate writing of output
* Join storm rate info to station details
* BUGFIX: Replace NaN with zeros
When converting to a feature class, the attributes with 'null' values in the first record were being stored as text fields, not double. This forces the NaN values to be zero, so eliminating "nuLL' strings in the GeoJSON.

* Code to join daily data to storm class