Develop (#8)
* change legend position and y axis color
* SystemError happens when import RocketClassifier
* openpyxl needed in classifyTimeSeries output code
* pandas need >=1.5.0 in convertStormCounts.py
* change the dir
* add a bit of comments at the head for new user
* Add a check for stray breakpoint calls
* WIP: Updating the classifier
* Fix long line in log message
* Add provenance to convert storm count data

Also clean up the classifyTimeSeries to make variable names a bit more intuitive

* Some more markdown discussion
* Fix pylint errors
* Clear output from notebook
* Change dat file name
* Extract accumulated rainfall

Extract accumulated precipitation in the hour preceding and the hour following the daily maximum gust event. Exclude any rainfall observations that have not passed BoM quality checks (i.e. only include obs where the rainfall quality flag is 'Y').

* Remove duplicate writing of output
* Join storm rate info to station details
* BUGFIX: Replace NaN with zeros

When converting to a feature class, the attributes with 'null' values in the first record were being stored as text fields, not double. This forces the NaN values to be zero, so eliminating "nuLL' strings in the GeoJSON.

* Code to join daily data to storm class
* Add weather codes to storm classifications
* Add cross tabulations
* Provenance info for cross tabulations
* Add Bayesian hierarchical model notebook
* Add a loop to evaluate performance of ROCKET for range of kernels
* Provenance updates
* Classifcation performance

Add a loop to fit the classification with varying numbers of random kernels, then calculate Heidke Skill Score and Gerrity Score, and repeat the calculations multiple times to report mean scores and timing estimates

* Tidy up file name variables
* Plot a selection of storms
* Pylint error fixes

---------

Co-authored-by: LiangHuLiu <liang.hu@ga.gov.au>